fix defered writes
The SerialTransport.write() method was not actually handling deferred
writes correctly. There were a number of issues with it.

1.  When a large amount of data was written, it would begin with a
    direct write to the serial port. Some, but not all of the data would
    be written. The remainder would be put in the _write_buffer and a
    writer would be added to the event loop.

    Once the buffer was drained, there was nothing to remove the writer
    from the event loop. This caused it to continually wake up and call
    _write_ready(), which would find no data to write and crash with an
    an AssertionError.

This is fixed by checking for no data in _write_ready() and removing the
writer from the event loop at that point.

Fixing that exposed the next issue.

2.  If data is written in chunks that exactly match the underlying
    serial port buffer size, then doing a direct write could deadlock.
    When the buffer is full, the write will raise an OSError with
    EWOULDBLOCK. PySerial catches this and retries the write
    indefinitely, hence the deadlock. This blocks the asyncio event
    loop from running at all.

To fix this issue, we register a poll object when the transport is
created. Then before doing a direct write, we poll the serial port to
make sure that writing will not block. If it would block, we skip the
direct write and go straight to buffering the data and adding the
writer to the event loop.

Fixing this exposed another issue.

3.  If the direct write was skipped, there was nothing to add the writer
    to the event loop. This caused the buffered data to never be sent.

This is fixed by calling _ensure_writer() when the direct write is
skipped.

A test is added that fails without these fixes and passes with them.
PyTest catches the AssertionError from 1 even though it happens in a
background task. Pytest (or the CI runner) should eventually time out
when 2 happens. An 3 triggers a TimeoutError in the test waiting for
one of the conditions to be met.